Implementa serviço de telemetria e ajusta tipos de dados
- Altera propriedades RPM, Marcha e Combustivel para float em TelemetryData para maior precisão.
- Adiciona e implementa TelemetryService, incluindo método ExibirTelemetria.
- Atualiza ITelemetryService para expor ExibirTelemetria.
- Modifica endpoint /telemetria para retornar dados do serviço.
- Inclui referências e atributo de plataforma no Program.cs.

+++ ACRH/Domain/Entities/TelemetryData.cs
@@ -6,9 +6,9 @@ namespace ACRH.Domain.Entities
     {
         public float Acelerador { get; set; }
         public float Freio { get; set; }
-        public int RPM { get; set; }
-        public int Marcha { get; set; }
-        public int Combustivel { get; set; }
+        public float RPM { get; set; }
+        public float Marcha { get; set; }
+        public float Combustivel { get; set; }
         public float[] Pneus { get; set; } = new float[4];
     }
 }
